fix: Synchronize processing of external changes
